[REFACTOR] Move onboarding modal to new directory structure
